Hago un picker generico
    refs:#288
